[#139] lint for check style
[#139] Test coverage ignoring
