+++ lib/constant/default-predefined-method-descriptor-registry.js
@@ -6,15 +6,15 @@
 
 'use strict'
 
-const MethodDescriptorBuilder2 = require('../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../context/method-descriptor-builder')
 const MethodType = require('./method-type')
 
 // DefaultPredefinedMethodDescriptorRegistry.java
 // Node Server Process SpanEvent.apiId predefined values
 class DefaultPredefinedMethodDescriptorRegistry {
   registryMethodDescriptor(apiMetaService) {
-    const nodeServerProcessBuilder = new MethodDescriptorBuilder2().setType(MethodType.WEB_REQUEST).setApiDescriptor('Node Server Process')
-    const asyncInvocationBuilder = new MethodDescriptorBuilder2().setType(MethodType.INVOCATION).setApiDescriptor('Asynchronous Invocation')
+    const nodeServerProcessBuilder = new MethodDescriptorBuilder().setType(MethodType.WEB_REQUEST).setApiDescriptor('Node Server Process')
+    const asyncInvocationBuilder = new MethodDescriptorBuilder().setType(MethodType.INVOCATION).setApiDescriptor('Asynchronous Invocation')
     this.nodeServerProcessDescriptor = apiMetaService.cacheApiWithBuilder(nodeServerProcessBuilder)
     this.asyncInvocationDescriptor = apiMetaService.cacheApiWithBuilder(asyncInvocationBuilder)
   }
