+++ lib/context/method-descriptor-builder.js
@@ -1,80 +1,27 @@
 /**
  * Pinpoint Node.js Agent
- * Copyright 2021-present NAVER Corp.
+ * Copyright 2023-present NAVER Corp.
  * Apache License v2.0
  */
 
 'use strict'
 
-const MethodDescriptor = require('./method-descriptor')
-
+const MethodDescriptor2 = require('../instrumentation/method-descriptor2')
 const MethodType = require('../constant/method-type')
 
 // https://v8.dev/docs/stack-trace-api#appendix%3A-stack-trace-format
 class MethodDescriptorBuilder {
-    constructor(moduleName, objectPath, methodName) {
-        this.moduleName = moduleName
-        this.objectPath = objectPath
-        this.methodName = methodName
+    constructor(functionName) {
         this.type = MethodType.DEFAULT
-        this.parameterDescriptor = '()'
         this.apiId = 0
-        this.computedApiDescriptor = this.makeApiDescriptor()
-        this.computedFullName = this.makeFullName(moduleName, objectPath)
+        this.parameterVariableNames = []
+        this.functionName = functionName
+        this.methodIndex = 2
+        this.locationFileNameIndex = 2
     }
 
-    makeApiDescriptor() {
-        const buffer = []
-        const typeFunctionString = this.typeFunction()
-        if (typeFunctionString && typeof typeFunctionString === 'string') {
-            buffer.push(typeFunctionString)
-        }
-        if (this.parameterDescriptor && typeof this.parameterDescriptor === 'string') {
-            buffer.push(this.parameterDescriptor)
-        }
-        return buffer.join('')
-    }
-
-    typeFunction() {
-        const buffer = []
-        if (this.moduleName && typeof this.moduleName === 'string') {
-            buffer.push(this.moduleName)
-        }
-        if (this.className && typeof this.className === 'string') {
-            buffer.push(this.className)
-        }
-        if (this.objectPath && typeof this.objectPath === 'string') {
-            buffer.push(this.objectPath)
-        }
-        return buffer.join('.')
-    }
-
-    static make(moduleName, stackRegexNamedGroup) {
-        if (!stackRegexNamedGroup || !stackRegexNamedGroup.functionName) {
-            return
-        }
-
-        const functionName = makeFunctionName(stackRegexNamedGroup)
-        const builder = new MethodDescriptorBuilder(moduleName, functionName, makeMethodName(stackRegexNamedGroup))
-            .setFunctionName(functionName)
-            .setLineNumber(makeLineNumber(stackRegexNamedGroup))
-            .setClassName(stackRegexNamedGroup.type)
-            .setFileName(stackRegexNamedGroup.fileName)
-
-        if (typeof stackRegexNamedGroup.location === 'string') {
-            builder.setLocation(`${stackRegexNamedGroup.location}${stackRegexNamedGroup.fileName}`)
-        } else {
-            builder.setLocation(`${stackRegexNamedGroup.fileName}`)
-        }
-        return builder
-    }
-
-    static makeWithNameGroup(namedGroup) {
-        return MethodDescriptorBuilder.make(undefined, namedGroup)
-    }
-
-    getModuleName() {
-        return this.moduleName
+    static makeRuntimeDetectionMethodDescriptorBuilder() {
+        return new MethodDescriptorBuilder().enableRuntimeDetection(true)
     }
 
     setType(type) {
@@ -82,8 +29,42 @@ class MethodDescriptorBuilder {
         return this
     }
 
-    setApiDescriptor(apiDescriptor) {
-        this.apiDescriptor = apiDescriptor
+    setClassName(className) {
+        this.className = className
+        return this
+    }
+
+    setLineNumber(lineNumber) {
+        this.lineNumber = lineNumber
+        return this
+    }
+
+    setLocation(location) {
+        this.location = location
+        return this
+    }
+
+    getFileName() {
+        return this.fileName
+    }
+
+    setFileName(fileName) {
+        this.fileName = fileName
+        return this
+    }
+
+    setParameterVariableNames(names) {
+        this.parameterVariableNames = names
+        return this
+    }
+
+    setFunctionName(functionName) {
+        this.functionName = functionName
+        return this
+    }
+
+    setMethodName(methodName) {
+        this.methodName = methodName
         return this
     }
 
@@ -92,147 +73,140 @@ class MethodDescriptorBuilder {
         return this
     }
 
-    setFullName(fullName) {
-        this.fullName = fullName
-        return this
-    }
-
-    getFullName() {
-        return this.fullName || this.computedFullName
-    }
-
-    setLineNumber(lineNumber) {
-        this.lineNumber = lineNumber
-        return this
-    }
-
-    getLineNumber() {
-        return this.lineNumber
-    }
-
-    setLocation(location) {
-        this.location = location
-        return this
-    }
-
-    getLocation() {
-        return this.location
-    }
-
-    setFileName(fileName) {
-        this.fileName = fileName
-        return this
-    }
-
-    getFileName() {
-        return this.fileName
-    }
-
-    build() {
-        const apiDescriptor = this.apiDescriptor || this.computedApiDescriptor
-        const fullName = this.fullName || this.computedFullName
-        return new MethodDescriptor(this.moduleName, this.objectPath, this.methodName, this.type, apiDescriptor, this.apiId, fullName, this.className, this.lineNumber, this.location)
-    }
-
-    makeFullNameWithApiDescriptor(moduleName, objectPath) {
-        if (moduleName && !objectPath) {
-            return `${moduleName}.<anonymous>`
-        }
-        return this.makeFullNameWithApiDescriptorAndParameterDescriptor()
-    }
-
-    makeFullNameWithApiDescriptorAndParameterDescriptor() {
-        let parameterDescriptor = this.parameterDescriptor
-        if (!parameterDescriptor || typeof parameterDescriptor !== 'string') {
-            parameterDescriptor = '()'
-        }
-        if (!this.moduleName && this.className && this.objectPath) {
-            return `${this.className}.${this.objectPath}${parameterDescriptor}`
-        }
-        return [this.objectPath, parameterDescriptor]
-            .filter(v => v)
-            .join('')
-    }
-
-    makeFullName(moduleName, objectPath) {
-        if (moduleName && !objectPath) {
-            return `${moduleName}.<anonymous>`
-        }
-
-        if (moduleName) {
-            return `${moduleName}.${this.makeFullNameWithApiDescriptor(moduleName, objectPath)}`
-        } else {
-            return `${this.makeFullNameWithApiDescriptor(moduleName, objectPath)}`
-        }
-    }
-
-    getFunctionName() {
-        return this.functionName
-    }
-
-    setFunctionName(functionName) {
-        this.functionName = functionName
-        return this
-    }
-
-    getClassName() {
-        return this.className
-    }
-
-    setClassName(className) {
-        this.className = className
-        this.computedApiDescriptor = this.makeApiDescriptor()
-        return this
-    }
-
-    setParameterDescriptor(parameterDescriptor) {
-        this.parameterDescriptor = parameterDescriptor
-        this.computedApiDescriptor = this.makeApiDescriptor()
-        this.computedFullName = this.makeFullName(this.moduleName, this.objectPath)
+    setApiDescriptor(apiDescriptor) {
+        this.apiDescriptor = apiDescriptor
         return this
     }
 
     getCacheId() {
-        return `${this.getFullName()}:${this.lineNumber}:${this.fileName}`
+        if (this.apiDescriptor) {
+            return this.apiDescriptor
+        }
+        
+        const cacheIds = [this.formattedStringOfCall()]
+
+        if (typeof this.fileName === 'string') {
+            cacheIds.push(this.fileName)
+        }
+
+        if (typeof this.lineNumber === 'number') {
+            cacheIds.push(this.lineNumber)
+        }
+        return cacheIds.join(':')
     }
 
-    static makeLineNumber(namedGroups) {
-        return makeLineNumber(namedGroups)
-    }
-}
-
-function makeFunctionName(namedGroup) {
-    const functionName = namedGroup.functionName
-    if (!functionName) {
-        return
+    toggleConstruct() {
+        this.construct = true
+        delete this.async
+        delete this.type
+        return this
     }
 
-    if (typeof functionName !== 'string') {
-        return
+    toggleAsync() {
+        this.async = true
+        delete this.construct
+        delete this.type
+        return this
     }
 
-    const computedGroups = functionName.match(/(?<functionName>.+)(?<=\.)<computed>$/)
-    if (computedGroups && computedGroups.length > 0 && namedGroup.methodName) {
-        return computedGroups.groups.functionName + namedGroup.methodName
-    }
-    if (functionName === '<computed>' && typeof namedGroup.type === 'string' && typeof namedGroup.methodName === 'string') {
-        return namedGroup.methodName
-    }
-    return namedGroup.functionName
-}
+    // https://v8.dev/docs/stack-trace-api#appendix%3A-stack-trace-format
+    formattedStringOfCall() {
+        if (this.needsMethodAndFunctionName() || (this.needsFunctionNameWhenConstructOrAsync())) {
+            return `<anonymous>`
+        }
 
-function makeMethodName(namedGroups) {
-    if (namedGroups.functionName == '<anonymous>' && !namedGroups.methodName && namedGroups.type) {
-        return namedGroups.functionName
-    }
-    return namedGroups.methodName || namedGroups.functionName
-}
+        if (this.construct) {
+            return `new ${this.functionName}`
+        } else if (this.async) {
+            return `async ${this.functionName}`
+        }
 
-function makeLineNumber(namedGroups) {
-    if (!namedGroups.lineNumber || typeof namedGroups.lineNumber !== 'string') {
-        return
+        const formatted = []
+        if (typeof this.className === 'string') {
+            formatted.push(`${this.className}.`)
+        }
+
+        if (typeof this.functionName === 'string') {
+            if (/<computed>$/.test(this.functionName) && typeof this.methodName === 'string') {
+                formatted.push(this.functionName.replace(/<computed>$/, this.methodName))
+                return formatted.join('')
+            }
+
+            formatted.push(this.functionName)
+        }
+
+        if (typeof this.methodName === 'string' && typeof this.functionName === 'string' && this.functionName !== this.methodName) {
+            formatted.push(` [as ${this.methodName}]`)
+        } else if (typeof this.methodName === 'string' && typeof this.functionName !== 'string') {
+            formatted.push(this.methodName)
+        }
+
+        return formatted.join('')
     }
-    return parseInt(namedGroups.lineNumber)
+
+    needsMethodAndFunctionName() {
+        return !this.functionName && !this.methodName
+    }
+
+    needsFunctionNameWhenConstructOrAsync() {
+        return (this.construct || this.async) && !this.functionName
+    }
+
+    isRuntimeDetection() {
+        return this.runtimeDetection
+    }
+
+    enableRuntimeDetection() {
+        this.runtimeDetection = true
+        return this
+    }
+
+    setMethodIndex(index) {
+        this.methodIndex = index
+        return this
+    }
+
+    setFileNameIndex(index) {
+        this.locationFileNameIndex = index
+        return this
+    }
+
+    makeCloneOf(functionName) {
+        let clone
+        if (functionName) {
+            clone = new MethodDescriptorBuilder(functionName)
+        } else {
+            clone = new MethodDescriptorBuilder(this.functionName)
+        }
+
+        clone.type = this.type
+        clone.parameterVariableNames = this.parameterVariableNames
+        clone.lineNumber = this.lineNumber
+        clone.fullName = this.fullName
+        clone.apiId = this.apiId
+        clone.apiDescriptor = this.apiDescriptor
+        clone.location = this.location
+        clone.className = this.className
+        return clone
+    }
+
+    build() {
+        const value = new MethodDescriptor2(this.functionName)
+        if (this.type !== 'Function' && typeof this.methodName === 'undefined') {
+            value.methodName = this.functionName
+        } else {
+            value.methodName = this.methodName
+        }
+        value.type = this.type
+        value.parameterVariableNames = this.parameterVariableNames
+        value.lineNumber = this.lineNumber
+        value.fullName = this.fullName
+        value.apiId = this.apiId
+        value.apiDescriptor = this.apiDescriptor || this.formattedStringOfCall()
+        value.location = this.location
+        value.className = this.className
+        return value
+    }
 }
 
 module.exports = MethodDescriptorBuilder
