+++ lib/instrumentation/call-stack.js
@@ -15,6 +15,10 @@ const callSite = (builder) => {
     Error.captureStackTrace(stackTrace, callSite)
     const callSites = arrayOfStackTrace(stackTrace)
 
+    return makeCloneBuilder(callSites, builder)
+}
+
+function makeCloneBuilder(callSites, builder) {
     let namedGroup = namedGroupLocationFileName(callSites, builder.locationFileNameIndex)
     if (builder.isRuntimeDetection()) {
         namedGroup = Object.assign(namedGroup, namedGroupTypeMethod(callSites, builder.methodIndex))
@@ -97,7 +101,7 @@ const makeLineNumber = (namedGroups) => {
 
 function namedGroupTypeMethod(callSites, index) {
     const callSite = callSiteOf(callSites, index)
-    const functionNameMatches = [typeFunctionNameMethodName, newFunctionName, asyncFunctionName]
+    const functionNameMatches = [newFunctionName, asyncFunctionName, typeFunctionNameMethodName]
     let group
     for (const match of functionNameMatches) {
         group = match(callSite)
@@ -148,5 +152,8 @@ const asyncFunctionName = (callSite) => {
 }
 
 module.exports = {
-    callSite
+    callSite,
+    namedGroupLocationFileName,
+    namedGroupTypeMethod,
+    makeCloneBuilder
 }
