+++ lib/instrumentation/context/annotation/annotations.js
@@ -6,12 +6,12 @@
 
 'use strict'
 
-const BooleanAnnotation = require("./boolean-annotation")
-const DataTypeAnnotation = require("./data-type-annotation")
-const NullAnnotation = require("./null-annotation")
-const IntAnnotation = require("./int-annotation")
-const ObjectAnnotation = require("./object-annotation")
-const StringAnnotation = require("./string-annotation")
+const BooleanAnnotation = require('./boolean-annotation')
+const DataTypeAnnotation = require('./data-type-annotation')
+const NullAnnotation = require('./null-annotation')
+const IntAnnotation = require('./int-annotation')
+const ObjectAnnotation = require('./object-annotation')
+const StringAnnotation = require('./string-annotation')
 
 class Annotations {
     static of(key, value) {
