+++ lib/instrumentation/module/express/express-route-interceptor.js
@@ -6,13 +6,13 @@
 
 'use strict'
 
-const MethodDescriptorBuilder2 = require("../../../context/method-descriptor-builder2")
-const patchLayer = require("./patch-layer")
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
+const patchLayer = require('./patch-layer')
 
 class ExpressRouteInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = MethodDescriptorBuilder2.makeRuntimeDetectionMethodDescriptorBuilder()
+        this.methodDescriptorBuilder = MethodDescriptorBuilder.makeRuntimeDetectionMethodDescriptorBuilder()
             .setClassName('Router')
             .setFileNameIndex(3)
     }
