+++ lib/instrumentation/module/express/express-use-interceptor.js
@@ -6,12 +6,12 @@
 
 'use strict'
 
-const MethodDescriptorBuilder2 = require("../../../context/method-descriptor-builder2")
-const patchLayer = require("./patch-layer")
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
+const patchLayer = require('./patch-layer')
 
 class ExpressUseInterceptor {
     constructor() {
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('use')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('use')
            .setClassName('Router')
            .setFileNameIndex(5)
     }
