+++ lib/instrumentation/module/express/patch-layer.js
@@ -6,9 +6,9 @@
 
 'use strict'
 
-const apiMetaService = require("../../../context/api-meta-service")
-const shimmer = require("../../../shimmer")
-const serviceType = require("./express-service-type")
+const apiMetaService = require('../../../context/api-meta-service')
+const shimmer = require('../../../shimmer')
+const serviceType = require('./express-service-type')
 const util = require('util')
 const errorReportedSymbol = Symbol('PinpointErrorReported')
