+++ lib/instrumentation/module/koa/koa-register-interceptor.js
@@ -7,13 +7,12 @@
 'use strict'
 
 const apiMetaService = require('../../../context/api-meta-service')
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
-const { Builder } = require('../../../utils/log/logger2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const serviceType = require('./koa-service-type')
 
 class KoaRegisterInterceptor {
     constructor() {
-        this.methodDescriptorBuilder = MethodDescriptorBuilder2.makeRuntimeDetectionMethodDescriptorBuilder()
+        this.methodDescriptorBuilder = MethodDescriptorBuilder.makeRuntimeDetectionMethodDescriptorBuilder()
             .setClassName('koa')
             .setFileNameIndex(3)
     }
