+++ lib/instrumentation/module/mongodb/mongo-json-data-annotation-key.js
@@ -6,6 +6,6 @@
 
 'use strict'
 
-const AnnotationKey = require("../../../context/annotation-key")
+const AnnotationKey = require('../../../context/annotation-key')
 
 module.exports = new AnnotationKey(150, 'MONGO-JSON-Data')
