+++ lib/instrumentation/module/mysql/mysql-cluster-of-interceptor.js
@@ -6,14 +6,14 @@
 
 'use strict'
 
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const InstrumentMethod = require('../../instrument-method')
 const MySQLGetConnectionInterceptor = require('./mysql-get-connection-interceptor')
 
 class MySQLClusterOfInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('of')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('of')
     }
 
     doInBeforeTrace(recorder, target, args) {
