+++ lib/instrumentation/module/mysql/mysql-connection-interceptor.js
@@ -7,10 +7,10 @@
 'use strict'
 
 const serviceType = require('./mysql-service-type')
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 class MySQLConnectionConnectInterceptor {
     constructor() {
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('connect')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('connect')
         this.methodDescriptorBuilder.setClassName('Connection')
         this.methodDescriptorBuilder.setParameterVariableNames('connectionUri')
     }
