+++ lib/instrumentation/module/mysql/mysql-create-connection-interceptor.js
@@ -9,7 +9,7 @@
 const serviceType = require('./mysql-service-type')
 const InstrumentMethod = require('../../instrument-method')
 const MySQLConnectionInterceptor = require('./mysql-connection-interceptor')
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const MySQLStatementExecuteQueryInterceptor = require('./mysql-statement-execute-query-interceptor')
 const databaseInfoSymbol = require('./mysql-database-information-symbol')
 const MySQLDatabaseInformationContext = require('./mysql-database-information-context')
@@ -17,7 +17,7 @@ const MySQLDatabaseInformationContext = require('./mysql-database-information-co
 class MySQLCreateConnectionInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('createConnection')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('createConnection')
         this.methodDescriptorBuilder.setParameterVariableNames('connectionUri')
     }
