+++ lib/instrumentation/module/mysql/mysql-create-pool-cluster-interceptor.js
@@ -6,14 +6,14 @@
 
 'use strict'
 
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const InstrumentMethod = require('../../instrument-method')
 const MySQLClusterOfInterceptor = require('./mysql-cluster-of-interceptor')
 
 class MySQLCreatePoolClusterInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('createPoolCluster')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('createPoolCluster')
     }
 
     doInBeforeTrace(recorder) {
