+++ lib/instrumentation/module/mysql/mysql-create-pool-interceptor.js
@@ -8,7 +8,7 @@
 
 const serviceType = require('./mysql-service-type')
 const InstrumentMethod = require('../../instrument-method')
-const MethodDescriptorBuilder2 = require("../../../context/method-descriptor-builder2")
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const databaseInfoSymbol = require('./mysql-database-information-symbol')
 const MySQLGetConnectionInterceptor = require('./mysql-get-connection-interceptor')
 const MySQLDatabaseInformationContext = require('./mysql-database-information-context')
@@ -16,7 +16,7 @@ const MySQLDatabaseInformationContext = require('./mysql-database-information-co
 class MySQLCreatePoolInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('createPool')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('createPool')
     }
 
     doInBeforeTrace(recorder, target, args) {
