+++ lib/instrumentation/module/mysql/mysql-get-connection-interceptor.js
@@ -7,7 +7,7 @@
 'use strict'
 
 const serviceType = require('./mysql-service-type')
-const MethodDescriptorBuilder2 = require("../../../context/method-descriptor-builder2")
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const databaseInfoSymbol = require('./mysql-database-information-symbol')
 const InstrumentMethod = require('../../instrument-method')
 const MySQLStatementExecuteQueryInterceptor = require('./mysql-statement-execute-query-interceptor')
@@ -15,7 +15,7 @@ const MySQLStatementExecuteQueryInterceptor = require('./mysql-statement-execute
 class MySQLGetConnectionInterceptor {
     constructor(traceContext) {
         this.traceContext = traceContext
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('getConnection')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('getConnection')
             .setClassName('Pool')
         this.serviceType = serviceType
     }
