+++ lib/instrumentation/module/mysql/mysql-statement-execute-query-interceptor.js
@@ -7,12 +7,12 @@
 'use strict'
 
 const serviceType = require('./mysql-execute-query-service-type')
-const MethodDescriptorBuilder2 = require('../../../context/method-descriptor-builder2')
+const MethodDescriptorBuilder = require('../../../context/method-descriptor-builder')
 const databaseInfoSymbol = require('./mysql-database-information-symbol')
 
 class MySQLStatementExecuteQueryInterceptor {
     constructor(className) {
-        this.methodDescriptorBuilder = new MethodDescriptorBuilder2('query')
+        this.methodDescriptorBuilder = new MethodDescriptorBuilder('query')
         this.methodDescriptorBuilder.setClassName(className)
         this.serviceType = serviceType
     }
