+++ lib/instrumentation/sql/sql-parser.js
@@ -6,7 +6,7 @@
 
 'use strict'
 
-const NormalizedSql = require("./normalized-sql")
+const NormalizedSql = require('./normalized-sql')
 
 // ref: DefaultSqlParser.java
 class SqlParser {
