+++ lib/utils/typed-buffer.js
@@ -6,7 +6,7 @@
 
 'use strict'
 
-const Int64BE = require("int64-buffer").Int64BE;
+const Int64BE = require('int64-buffer').Int64BE;
 
 class TypedBuffer {
   constructor(buffer, size) {
